Move `Checker` class to its own module
