+++ src/checkers/chromiumchecker.py
@@ -7,7 +7,6 @@ import aiohttp
 
 from ..lib import NETWORK_ERRORS
 from ..lib.externaldata import (
-    Checker,
     ExternalBase,
     ExternalData,
     ExternalGitRepo,
@@ -15,6 +14,7 @@ from ..lib.externaldata import (
 )
 from ..lib.utils import get_extra_data_info_from_url
 from ..lib.errors import CheckerMetadataError, CheckerFetchError
+from ..lib.checkers import Checker
 
 log = logging.getLogger(__name__)
