+++ src/checkers/gitchecker.py
@@ -5,9 +5,10 @@ from distutils.version import LooseVersion
 
 import semver
 
-from ..lib.externaldata import Checker, ExternalBase, ExternalGitRepo, ExternalGitRef
+from ..lib.externaldata import ExternalBase, ExternalGitRepo, ExternalGitRef
 from ..lib.utils import git_ls_remote
 from ..lib.errors import CheckerQueryError, CheckerFetchError
+from ..lib.checkers import Checker
 
 log = logging.getLogger(__name__)
