+++ src/checkers/jetbrainschecker.py
@@ -1,8 +1,9 @@
 import datetime
 import logging
 
-from ..lib.externaldata import ExternalBase, ExternalFile, Checker
+from ..lib.externaldata import ExternalBase, ExternalFile
 from ..lib.checksums import MultiDigest
+from ..lib.checkers import Checker
 
 log = logging.getLogger(__name__)
