+++ tests/test_checker.py
@@ -30,7 +30,8 @@ from xml.dom import minidom
 import aiohttp
 
 from src.lib.utils import init_logging
-from src.lib.externaldata import ExternalData, Checker
+from src.lib.externaldata import ExternalData
+from src.lib.checkers import Checker
 from src.lib.checksums import MultiDigest
 from src.lib.errors import CheckerFetchError
 from src.manifest import ManifestChecker
